requirements.txt
opencv-python == 4.11.0.86
Pillow == 11.2.1
numpy == 2.1.3